// lib/schemas/auth.ts
import { z } from "zod";

export const loginSchema = z.object({
  email: z.string().refine((val) => val === 'admin' || z.string().email().safeParse(val).success, {
    message: "Invalid email address",
  }),
  password: z.string().min(4, "Password must be at least 4 characters"),
});

export const signupSchema = z.object({
  fullName: z.string().min(2, "Full name must be at least 2 characters"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});